Decompose token response parsing in preparation for refresh responses
In this pass, we assume the refresh response will not have an id token
(they're optional, we'll deal with them in a later pass).

This commit creates a new token parsing function that will merge the
existing id token with new access and refresh tokens, as well as a new
access token expiry.

Signed-off-by: Tyler Schultz <tschultz@pivotal.io>