Fix flaky file watcher test (#258)
* Fix falky file watcher test

The test method `waitForRead` was sometimes signaling due to accumulated
reads, even it was inteded to wait for a read after a new content was
set in the mock file.

This fixes it by waiting to singal read happened when the read content
is the same as the expected one.

Signed-off-by: Sergi Castro <sergi@tetrate.io>

* Add more specific methods

Signed-off-by: Sergi Castro <sergi@tetrate.io>

---------

Signed-off-by: Sergi Castro <sergi@tetrate.io>