Add options to configure TLS and mTLS connections to Redis (#294)
This change allows configuring TLS properties for the redis connection
as well as the password. The values can be set in plain text in the
configuration or as references to files.

Note that even though the content of the Redis certificates or password
may live in a secret, I think it is more appropriate to reference them
as a file, not the secret, because:

* If the secret is mounted in the Authservice, the contents are ready
since the Authservice start, removing any eventual consistency until the
k8s reconciler has picked up the secret and propagated the updated
value.
* It does not require the AuthService to have additional permissions to
read secrets in the cluster.

The followign example shows a configuration for Redis using mTLS and
credentials, mounted from secrets:
```yaml
"redis_session_store_config": {
  "server_uri": "redis://redis.redis.svc.cluster.local:6379",
  "username": "authservice",
  "password_file": "/etc/authservice/redis/password",
  "tls_config": {
    "trusted_ca_file": "/etc/authservice/certs/ca.crt",
    "client_cert_file": "/etc/authservice/certs/tls.crt",
    "client_key_file": "/etc/authservice/certs/tls.key"
  }
},
```

For the Authservice use cases, mounting the secrets is what makes more
sense, and I'll propose deprecating the secret reference APIs in favor
of local files.

cc @ogarrett @shamusx

---------

Signed-off-by: Ignasi Barrera <ignasi@tetrate.io>